# algohive/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.AuthResponse:
    properties:
      email:
        type: string
      firstname:
        type: string
      groups:
        items:
          $ref: '#/definitions/models.Group'
        type: array
      last_connected:
        type: string
      lastname:
        type: string
      permissions:
        type: integer
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      token:
        type: string
      user_id:
        type: string
    type: object
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.RegisterRequest:
    properties:
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - firstname
    - lastname
    - password
    type: object
...
  models.Catalog:
    properties:
      address:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      scopes:
        items:
          $ref: '#/definitions/models.Scope'
        type: array
    type: object
  models.Competition:
    properties:
      api_environment:
        $ref: '#/definitions/models.Catalog'
      api_environment_id:
        type: string
      api_theme:
        type: string
      description:
        type: string
      finished:
        type: boolean
      groups:
        items:
          $ref: '#/definitions/models.Group'
        type: array
      id:
        type: string
      show:
        type: boolean
      title:
        type: string
      tries:
        items:
          $ref: '#/definitions/models.Try'
        type: array
    type: object
  models.Group:
    properties:
      competitions:
        items:
          $ref: '#/definitions/models.Competition'
        type: array
      description:
        type: string
      id:
        type: string
      name:
        type: string
      scope_id:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.Role:
    properties:
      id:
        type: string
      name:
        type: string
      permissions:
        type: integer
      scopes:
        items:
          $ref: '#/definitions/models.Scope'
        type: array
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.Scope:
    properties:
      catalogs:
        items:
          $ref: '#/definitions/models.Catalog'
        type: array
      description:
        type: string
      groups:
        items:
          $ref: '#/definitions/models.Group'
        type: array
      id:
        type: string
      name:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
    type: object
  models.Try:
    properties:
      attempts:
        type: integer
      competition_id:
        type: string
      end_time:
        type: string
      id:
        type: string
      puzzle_id:
        type: string
      puzzle_index:
        type: integer
      puzzle_lvl:
        type: string
      score:
        type: number
      start_time:
        type: string
      step:
        type: integer
      user_id:
        type: string
    type: object
  models.User:
    properties:
      blocked:
        type: boolean
      email:
        type: string
      firstname:
        type: string
      groups:
        items:
          $ref: '#/definitions/models.Group'
        type: array
      id:
        type: string
      last_connected:
        type: string
      lastname:
        type: string
      password:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
    type: object
...
info:
  contact:
    email: ericphlpp@proton.me
    name: AlgoHive Support
  description: This is the API documentation for the AlgoHive API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Swagger AlgoHive API
  version: 1.0.0
paths:
  /auth/check:
    get:
      consumes:
      - application/json
      description: Check if the sent cookie token session is still valid and return
        user data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Check if the sent cookie token session is still valid and return user
        data
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token
      parameters:
      - description: Login Credentials
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User Login
      tags:
      - Auth
  /auth/logout:
    post:
      description: Logout a user by invalidating their token
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: User Logout
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: Registration Details
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User Register
      tags:
      - Auth
...
  /competitions/{id}/groups:
    get:
      consumes:
      - application/json
      description: Get all groups that have access to the specified competition
      parameters:
      - description: Competition ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Group'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get all groups with access to a competition
      tags:
      - Competitions
  /competitions/{id}/groups/{group_id}:
    delete:
      consumes:
      - application/json
      description: Remove group access to a competition
      parameters:
      - description: Competition ID
        in: path
        name: id
        required: true
        type: string
      - description: Group ID
        in: path
        name: group_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Competition'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Remove a group from a competition
      tags:
      - Competitions
    post:
      consumes:
      - application/json
      description: Allow users in a group to access a competition
      parameters:
      - description: Competition ID
        in: path
        name: id
        required: true
        type: string
      - description: Group ID
        in: path
        name: group_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Competition'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Add a group to a competition
      tags:
      - Competitions
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"